Authorizer will use authentication jaas creds.
Previously, username and password were pulled from the JAAS
configuration as defined by the JVM JAAS utilities. However,
in cases where the JAAS configuration was defined in the Kafka
properties file, the JVM JAAS context was not set directly, and so the
username and password could not be found.

This change simply allows the JAAS context as passed to the
SimpleLDAPAuthentication.configure method to be cached, and then reused
for the authorization code.